requirements.txt
langchain
openai
streamlit
python-dotenv
PyMuPDF
faiss-cpu